requirements.txt
python-telegram-bot==20.6
python-dotenv
python-telegram-bot[webhooks]
requests
